Analyzing OLAP cube stored in Microsoft SQL Server Analysis Services(SSAS) from web application using XML for Analysis.
List of schemas and cubes are dynamically generated based on SSAS
server name passed. Discover command in XML for Analysis are wrapped in
SOAP envelop and sent to SSAS; the XML response is parsed and data is
extracted to build a TreeView containing all dimensions and measures in
the cube.
User issues an MDX query which is converted to Execute command in XML
for Analysis and sent to SSAS, the response is parsed and result is
used to populate chart.